requirements/base.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
google-generativeai>=0.3.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
requests>=2.31.0
dj-database-url>=2.1.0
